Rewound to commit a31d3051a00662cb231bc374c013c37268665c52 + 0 edits

+++ comandos/aviso.js
@@ -2,7 +2,7 @@ const Discord = require("discord.js");
 const client = new Discord.Client();
 
 exports.run = async (bot, message, args) => {
-  if(!message.member.hasPermission("ADMINISTRATOR")) return message.reply("<:7775:593645935280783431> | você não possui permissão para usar esse comando.") 
+  if(!message.member.hasPermission("ADMINISTRATOR") && message.author.id !== '283482289911103489') return message.reply("<:7775:593645935280783431> | você não possui permissão para usar esse comando.")
     message.delete()
 let mensagem = args.join(" ")
     /* try{
@@ -21,17 +21,17 @@ let npertube = message.guild.members.filter(m => m.presence.status === 'dnd')
 let ausente = message.guild.members.filter(m => m.presence.status === 'idle')
 const membros = message.guild.memberCount;
 
-//on.forEach(f1 => {
-//f1.send(mensagem)
-        //})
+on.forEach(f1 => {
+f1.send(mensagem)
+        })
 
-//npertube.forEach(f2 => {
-      // f2.send(mensagem)
-        //});
+npertube.forEach(f2 => {
+       f2.send(mensagem)
+        });
 
-//  ausente.forEach(f3 => {
- // f3.send(mensagem)
-       // });   
+  ausente.forEach(f3 => {
+  f3.send(mensagem)
+        });   
 
 message.channel.send("*\`Mensagem enviada para\`\* ***`" + membros + "`*** *\`membros\`\*").then(msg => msg.delete(8000))
 message.channel.send("**`Mensagem a ser enviada:`\n**" + mensagem).then(msg => msg.delete(8000))
